Project Structure
I'm looking into setting the dev environment up with Vite, but I'm not totally quite sure how to do that yet. This repo seems to be close to what I may be looking for. I don't like templates because I like to understand how the code I'm using works. So installing a command to just 'do it' for me doesn't teach me anything about how they set it up, it just sets it up for me. This one from the first few glances appears to be more user-centric though, so I think I might be able to extrapolate more things from how it works with it. It's also great that it uses Svelte here too, I'm also curious about how that would play into things! (Well, that's not that fancy with Vite, that's already something I've been testing with. But, it's new for in combination with Electron :) )

https://www.reddit.com/r/electronjs/comments/13svjn0/vite_svelte_electron_tailwind/
https://github.com/feernandobraga/vitesvelctron

Yeah so this is just organizing things a little more like a regular web dev project, as it was a bit hectic to start with originally. This looks more like a dedicated Vite app now, at least almost hehe.

I took inspiration from the linked repo, and Tauri's folder structure, in regards to the `electron` folder. I do like seeing the 'backend' app code in a separate folder, that seems to help with splitting things up in your mind as to what is implemented where, and what can import from what. I like it! Now that I have TypeScript, I think I can use types to type out the Context Bridge behavior for Electron, that would be very nice for message passing between the two contexts. I wonder if there's a TRPC-ish thing like that, for Electron? I want to do something like that with Web Workers too for multithreading. Also related/unrelated to that, I think NBT could be an interesting way to pass data between contexts, at least in the case of Flatlands, where game objects might make sense to be serialized to buffers along the way.