Strict Types
Was going to look into making a script to make Dark Mode-like images for myself, using an implementation of the Canvas API in Node, but I didn't like the need for a separate implementation for that. I also didn't want to build the whole thing in the browser, so thought it might be nice to build it in Electron, allowing it to be local and still browser-ish. I'm still not fully on-board with making Electron apps, even with how capable they are. I'm too much into PWAs, as I like how much more they work on different platforms. Was setting up a project for that, then I remembered I should probably just use how I did this project rather than figuring out how to implement it all from scratch again. Decided to end up just updating this one to have the latest things I meant to continue with it, then work on that one. Might come back to it, not sure. Already coded a lot today, so I think my creative capacity is running low, needs rest :)